+++ modules/micro-service/outputs.tf
@@ -1,4 +1,7 @@
-
+output "name" {
+  description = "The name of the service."
+  value       = local.service.name
+}
 output "configurations" {
   value       = module.configurations
   description = "The configurations object."
@@ -12,4 +15,38 @@ output "release_id" {
 output "volumes" {
   value       = local.volumes
   description = "The list of volumes mounted to the service."
+}
+
+output "namespace" {
+  description = "The namespace within kubernetes this service is in."
+  value       = local.namespace
+}
+
+output "domain" {
+  description = "The fully qualified domain name of the service."
+  value       = local.service.domain
+}
+
+output "parent_domain" {
+  description = "The parent domain of the service."
+  value       = local.service.parent_domain
+}
+
+output "image" {
+  description = "The actual image in use atm."
+  value       = local.service.docker_image
+}
+
+## leave this commented out, it's useful when deciding what to output
+# output "service" {
+#   description = "The output of the service from the provider."
+#   value       = local.service
+# }
+
+output "release_header" {
+  description = "This is an object with name and value. The name is the header name and the value is header value which is the value of the release_id."
+  value = var.lb.release_header ? {
+    name  = local.release_header_name
+    value = local.release_id
+  } : null
 }